Proyecto1/Aws/requirements.txt
aws-cdk-lib==2.147.0
constructs
